Move kjobctl-docs to cmd and kjobctl to separate folder
