+++ cmd/experimental/kjobctl/Makefile
@@ -59,7 +59,7 @@ artifacts: kustomize ## Generate release artifacts.
 	if [ -d artifacts ]; then rm -rf artifacts; fi
 	mkdir -p artifacts
 	$(KUSTOMIZE) build config/default -o artifacts/manifests.yaml
-	CGO_ENABLED=$(CGO_ENABLED) GO_CMD="$(GO_CMD)" LD_FLAGS="$(LD_FLAGS)" BUILD_DIR="artifacts" BUILD_NAME=kubectl-kjob PLATFORMS="$(CLI_PLATFORMS)" ./hack/multiplatform-build.sh ./cmd/main.go
+	CGO_ENABLED=$(CGO_ENABLED) GO_CMD="$(GO_CMD)" LD_FLAGS="$(LD_FLAGS)" BUILD_DIR="artifacts" BUILD_NAME=kubectl-kjob PLATFORMS="$(CLI_PLATFORMS)" ./hack/multiplatform-build.sh ./cmd/kjobctl/main.go
 
 ##@ Development
 
@@ -152,19 +152,17 @@ uninstall: manifests kustomize ## Uninstall CRDs from the K8s cluster specified 
 
 .PHONY: kubectl-kjob
 kubectl-kjob: embeded-manifest ## Build kubectl-kjob binary file to bin directory.
-	CGO_ENABLED=$(CGO_ENABLED) $(GO_BUILD_ENV) $(GO_CMD) build -ldflags="$(LD_FLAGS)" -o bin/kubectl-kjob cmd/main.go
+	CGO_ENABLED=$(CGO_ENABLED) $(GO_BUILD_ENV) $(GO_CMD) build -ldflags="$(LD_FLAGS)" -o bin/kubectl-kjob cmd/kjobctl/main.go
 
 .PHONY: kjobctl-docs
 kjobctl-docs: ## Build kjobctl-docs binary file to bin directory.
-	cd $(PROJECT_DIR)/hack/tools && \
-	go mod download && \
-	$(GO_BUILD_ENV) $(GO_CMD) build -ldflags="$(LD_FLAGS)" -o $(PROJECT_DIR)/bin/kjobctl-docs ./kjobctl-docs/main.go
+	$(GO_BUILD_ENV) $(GO_CMD) build -ldflags="$(LD_FLAGS)" -o $(PROJECT_DIR)/bin/kjobctl-docs ./cmd/kjobctl-docs/main.go
 
 .PHONY: generate-kjobctl-docs
 generate-kjobctl-docs: kjobctl-docs ## Generate kjobctl docs.
 	rm -Rf $(PROJECT_DIR)/docs/commands/*
 	$(PROJECT_DIR)/bin/kjobctl-docs \
-		$(PROJECT_DIR)/hack/tools/kjobctl-docs/templates \
+		$(PROJECT_DIR)/cmd/kjobctl-docs/templates \
 		$(PROJECT_DIR)/docs/commands
 
 ##@ Dependencies
