+++ cmd/experimental/kjobctl/hack/tools/go.mod
@@ -16,8 +16,6 @@ require (
 	sigs.k8s.io/kustomize/kustomize/v5 v5.4.3
 )
 
-replace sigs.k8s.io/kueue/cmd/experimental/kjobctl => ../..
-
 require (
 	github.com/Azure/go-ansiterm v0.0.0-20210617225240-d185dfc1b5a1 // indirect
 	github.com/MakeNowJust/heredoc v1.0.0 // indirect
